requirements.txt
Faker==18.7.0
matplotlib==3.7.1
numpy==1.22.3
onnx==1.14.0
onnxruntime==1.14.1
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
opencv_python_headless==4.6.0.66
Pillow==9.0.1
Requests==2.31.0
scikit_learn==1.1.2
torch==1.12.0
torchsummary==1.5.1
torchvision==0.13.0
